// src/app/enums/company-segment.enum.ts
export enum CompanySegmentEnum {
  //Agribusiness
  bovine = 'bovine',
  pork = 'pork',
  birds = 'birds',
  other_meat_animals = 'other_meat_animals',
  dairy = 'dairy',
  soy = 'soy',
  wood = 'wood',
  sugar_cane = 'sugar_cane',
  other_platations = 'other_platations',
  subsistence_plantation = 'subsistence_plantation',

  //Industry
  heavy_industry = 'heavy_industry',
  construction = 'construction',
  automotive = 'automotive',
  food = 'food',
  textile = 'textile',
  pharmaceutical = 'pharmaceutical',
  technology = 'technology',
  energy = 'energy',
  another_transformation = 'another_transformation',
  electronics = 'electronics',
  
  //Services
  commerce_food = 'commerce_food',
  commerce_clothing = 'commerce_clothing',
  commerce_pharmacy = 'commerce_pharmacy',
  education = 'education',
  accommodation = 'accommodation',
  cleaning = 'cleaning',
  transport = 'transport',
  sports = 'sports',
  leisure = 'leisure',
  consultancy = 'consultancy',
  resale = 'resale',
  miscellaneous = 'miscellaneous',
  liberal_profession = 'liberal_profession'
}